fix: Address PR review feedback for profile error handling
- Move ProfileTest to JVM-specific location (.jvm/src/test/scala) since it uses Java NIO
- Simplify environment variable replacement logic for better readability
- Add helper method withMockHome to reduce test duplication
- Fix missing sys.props restoration in test cleanup
